Allow blank soap actions on operations
Soap actions are allowed to be blank when parsing a WSDL
when the optional parameter `allow_blank_soap_action: true`
is passed as an option when calling `Soap.init_model/3`.
Additionally, a test with fixture is included to ensure the
behavior of the option is honored when the param is included
and omitted.